Rename test classes for consistency: append 'Tests' to class names
