+++ tests/Pipaslot.Mediator.Tests/E2E/ResultTypes/NotNullBooleanTests.cs
@@ -1,17 +1,16 @@
 ﻿using Pipaslot.Mediator.Abstractions;
-using System;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Pipaslot.Mediator.Tests.E2E.ResultTypes;
 
-public class NotNullDateTime
+public class NotNullBooleanTests
 {
     [Theory]
-    [InlineData("2021-09-01")]
-    public async Task Execute_ShouldPass(string dateString)
+    [InlineData(true)]
+    [InlineData(false)]
+    public async Task Execute_ShouldPass(bool value)
     {
-        var value = DateTime.Parse(dateString);
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
         var result = await sut.Execute(new FakeAction(value));
         Assert.True(result.Success);
@@ -19,20 +18,20 @@ public class NotNullDateTime
     }
 
     [Theory]
-    [InlineData("2021-09-01")]
-    public async Task ExecuteUnhandled_ShouldPass(string dateString)
+    [InlineData(true)]
+    [InlineData(false)]
+    public async Task ExecuteUnhandled_ShouldPass(bool value)
     {
-        var value = DateTime.Parse(dateString);
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
         var result = await sut.ExecuteUnhandled(new FakeAction(value));
         Assert.Equal(value, result);
     }
 
-    public record FakeAction(DateTime Value) : IMediatorAction<DateTime>;
+    public record FakeAction(bool Value) : IMediatorAction<bool>;
 
-    public class FakeActionHandler : IMediatorHandler<FakeAction, DateTime>
+    public class FakeActionHandler : IMediatorHandler<FakeAction, bool>
     {
-        public Task<DateTime> Handle(FakeAction action, CancellationToken cancellationToken)
+        public Task<bool> Handle(FakeAction action, CancellationToken cancellationToken)
         {
             return Task.FromResult(action.Value);
         }
