+++ tests/Pipaslot.Mediator.Tests/E2E/ResultTypes/NotNullDateTimeTests.cs
@@ -1,17 +1,17 @@
 ﻿using Pipaslot.Mediator.Abstractions;
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Pipaslot.Mediator.Tests.E2E.ResultTypes;
 
-public class NotNullDecimal
+public class NotNullDateTimeTests
 {
     [Theory]
-    [InlineData(-100)]
-    [InlineData(0)]
-    [InlineData(100)]
-    public async Task Execute_ShouldPass(decimal value)
+    [InlineData("2021-09-01")]
+    public async Task Execute_ShouldPass(string dateString)
     {
+        var value = DateTime.Parse(dateString);
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
         var result = await sut.Execute(new FakeAction(value));
         Assert.True(result.Success);
@@ -19,21 +19,20 @@ public class NotNullDecimal
     }
 
     [Theory]
-    [InlineData(-100)]
-    [InlineData(0)]
-    [InlineData(100)]
-    public async Task ExecuteUnhandled_ShouldPass(decimal value)
+    [InlineData("2021-09-01")]
+    public async Task ExecuteUnhandled_ShouldPass(string dateString)
     {
+        var value = DateTime.Parse(dateString);
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
         var result = await sut.ExecuteUnhandled(new FakeAction(value));
         Assert.Equal(value, result);
     }
 
-    public record FakeAction(decimal Value) : IMediatorAction<decimal>;
+    public record FakeAction(DateTime Value) : IMediatorAction<DateTime>;
 
-    public class FakeActionHandler : IMediatorHandler<FakeAction, decimal>
+    public class FakeActionHandler : IMediatorHandler<FakeAction, DateTime>
     {
-        public Task<decimal> Handle(FakeAction action, CancellationToken cancellationToken)
+        public Task<DateTime> Handle(FakeAction action, CancellationToken cancellationToken)
         {
             return Task.FromResult(action.Value);
         }
