+++ tests/Pipaslot.Mediator.Tests/E2E/ResultTypes/NotNullObjectTests.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace Pipaslot.Mediator.Tests.E2E.ResultTypes;
 
-public class NullableObject
+public class NotNullObjectTests
 {
     [Fact]
     public async Task Execute_ReturnsValue_ShouldPass()
@@ -16,12 +16,13 @@ public class NullableObject
     }
 
     [Fact]
-    public async Task Execute_ReturnsNull_ShouldPass()
+    public async Task Execute_ReturnsNull_ShouldFailButSuccess()
     {
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
-        var result = await sut.Execute(new FakeAction(true));
-        Assert.True(result.Success, result.GetErrorMessage());
-        Assert.Null(result.Result);
+        var action = new FakeAction(true);
+        var result = await sut.Execute(action);
+        Assert.True(result
+            .Success); // Would be nice to get false and detect if null was provided when null is not expected, but we can not achieve it in the current C# version
     }
 
     [Fact]
@@ -33,23 +34,24 @@ public class NullableObject
     }
 
     [Fact]
-    public async Task ExecuteUnhandled_ReturnsNull_ShouldPass()
+    public async Task ExecuteUnhandled_ReturnsNull_ShouldFailButSuccess()
     {
+        var action = new FakeAction(true);
+        // Would be nice to get failure and detect if null was provided when null is not expected, but we can not achieve it in the current C# version
         var sut = Factory.CreateMediatorWithHandlers<FakeActionHandler>();
-        var result = await sut.ExecuteUnhandled(new FakeAction(true));
-        Assert.Null(result);
+        await sut.ExecuteUnhandled(action);
     }
 
-    public record FakeAction(bool ReturnNull) : IMediatorAction<FakeResult?>;
+    public record FakeAction(bool ReturnNull) : IMediatorAction<FakeResult>;
 
     public record FakeResult;
 
-    public class FakeActionHandler : IMediatorHandler<FakeAction, FakeResult?>
+    public class FakeActionHandler : IMediatorHandler<FakeAction, FakeResult>
     {
-        public Task<FakeResult?> Handle(FakeAction action, CancellationToken cancellationToken)
+        public Task<FakeResult> Handle(FakeAction action, CancellationToken cancellationToken)
         {
             var result = action.ReturnNull ? null : new FakeResult();
-            return Task.FromResult(result);
+            return Task.FromResult(result!);
         }
     }
 }
