+++ tests/Pipaslot.Mediator.Tests/Mediator_ExecutionExceptionSubtypeTests.cs
@@ -12,8 +12,8 @@ namespace Pipaslot.Mediator.Tests;
 /// that apply across all three at once - existing <c>catch (MediatorExecutionException)</c> code still works, and
 /// <c>Response</c> is populated the same way regardless of subtype - so they don't belong to any single trigger's
 /// own test file. Each trigger's own "throws the correct subtype" assertion lives with that trigger's test
-/// (no-handler in E2E.Nohandler, status-Failed-without-throw in E2E.NoHandlerWithoutError, missing-result in
-/// E2E.ResultWasTakenFromTheContext) - don't duplicate that here.
+/// (no-handler in E2E.NoHandlerTests, status-Failed-without-throw in E2E.NoHandlerWithoutErrorTests, missing-result in
+/// E2E.ResultWasTakenFromTheContextTests) - don't duplicate that here.
 /// </summary>
 public class Mediator_ExecutionExceptionSubtypeTests
 {
@@ -25,8 +25,8 @@ public class Mediator_ExecutionExceptionSubtypeTests
     {
         var sut = trigger switch
         {
-            SubtypeTrigger.MissingResult => Factory.CreateConfiguredMediator(c => c.Use<E2E.ResultWasTakenFromTheContext.RemoveResultFromContextMiddleware>()),
-            SubtypeTrigger.UnhandledError => Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerWithoutError.BlockRequestMiddleware>()),
+            SubtypeTrigger.MissingResult => Factory.CreateConfiguredMediator(c => c.Use<E2E.ResultWasTakenFromTheContextTests.RemoveResultFromContextMiddleware>()),
+            SubtypeTrigger.UnhandledError => Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerWithoutErrorTests.BlockRequestMiddleware>()),
             _ => Factory.CreateConfiguredMediator(),
         };
 
@@ -35,7 +35,7 @@ public class Mediator_ExecutionExceptionSubtypeTests
         {
             if (trigger == SubtypeTrigger.UnhandledError)
             {
-                await sut.ExecuteUnhandled(new E2E.NoHandlerWithoutError.BlockedRequest());
+                await sut.ExecuteUnhandled(new E2E.NoHandlerWithoutErrorTests.BlockedRequest());
             }
             else
             {
@@ -71,13 +71,13 @@ public class Mediator_ExecutionExceptionSubtypeTests
         // Response is built by the MediatorExecutionException base constructor, unchanged by this unit - this test
         // proves that reading Response off the now-thrown MediatorUnhandledErrorException still reflects the real
         // pipeline results exactly as it did off the base type before this change (no regression in what a caught
-        // exception exposes). Reuses NoHandlerAndErrorReturned's fixture rather than a local copy.
-        var sut = Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerAndErrorReturned.AddErrorAndEndMiddleware>());
+        // exception exposes). Reuses NoHandlerAndErrorReturnedTests' fixture rather than a local copy.
+        var sut = Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerAndErrorReturnedTests.AddErrorAndEndMiddleware>());
         var action = new SingleHandler.Request(true);
 
         var ex = await Assert.ThrowsAsync<MediatorUnhandledErrorException>(() => sut.ExecuteUnhandled(action));
 
         Assert.False(ex.Response.Success);
-        Assert.Contains(ex.Response.Results.OfType<Notification>(), n => n.Content == E2E.NoHandlerAndErrorReturned.Error);
+        Assert.Contains(ex.Response.Results.OfType<Notification>(), n => n.Content == E2E.NoHandlerAndErrorReturnedTests.Error);
     }
 }
